Fix incognito notification/cleanup for android multiwindow.
When moving an NTP (very specific to the NTP URL) across
CTA instances, it would put the ref in AsyncTabParams, but
would not use it when creating the tab in the second window.

This was bad because it kept a reference to the incognito
tab forever (preventing the profile from being deleted and
the notification from going away).

This also changes the PendingIntent flag to allow the
notification to take an action as long as it is visible
(then if cleaning up the incognito tabs fails the first
time, the second click might succeed).

BUG=673582, 673489

Review-Url: https://codereview.chromium.org/2574143004
Review-Url: https://codereview.chromium.org/2603793002 .
Cr-Original-Commit-Position: refs/heads/master@{#438648}
Cr-Commit-Position: refs/branch-heads/2924@{#620}
Cr-Branched-From: 20dccad4a0c728c2e904bd3cea6523f4fb2eea0c [formerly 3a87aecc31cd1ffe751dd72c04e5a96a1fc8108a]-refs/heads/master@{#433059}

Former-commit-id: 9af8d43fec1953c20f6804c0f80dfe6ea0da27f7