Just clamp to limit-epsilon.
I think the -0.5f was an implementation detail of Herb's bilerp that we
don't need here.  It happened to also be clamping us to something less
than limit (limit-0.5), so we do need to replace that with a little
nudge to keep us on tile.

Change-Id: I4ebd32e0ad38c724a17dc8bc35d9ea228eeeca32
Reviewed-on: https://skia-review.googlesource.com/7338
Commit-Queue: Mike Klein <mtklein@chromium.org>
Reviewed-by: Herb Derby <herb@google.com>
Reviewed-by: Florin Malita <fmalita@chromium.org>

Former-commit-id: abddea0d8c05103dfb3ea63744f178b4ef4075c9