"can we?" -> "do we want to?" for SkRasterPipelineBlitter
There has been a supported() function in SkRasterPipelineBlitter.cpp for
a long time that's becoming increasingly misnamed.  That blitter ought
to be able to handle all destination formats.

This CL moves that logic outside to the creator of the blitter, changing
it from "can we handle this format?" to "do we want to use this blitter
for this format?".

In other CLs I'm working to make creating a pipeline blitter never fail.

Change-Id: Ie59fb8ec6e63d215d1baef439e464e8f0ab3ae4d
Reviewed-on: https://skia-review.googlesource.com/15842
Reviewed-by: Mike Reed <reed@google.com>
Commit-Queue: Mike Klein <mtklein@chromium.org>

Former-commit-id: d1835d7491223f94b07d48e5af8d29fb6742bf4e