Experimental Go bindings for Skia
This CL implements Go bindings for a subset of the functions in the C API. 
It implements a Go version of the C demo program in 
experimental/c-api-example/skia-c-example.c and the output is identical. 
(Checked by hand). 

The main purpose is to establish a pattern of calling the Skia C API that 
is memory safe and provides a idiomatic Go interface to Skia. 

Follow up CLs will cover the entire C API, add documentation and establish 
a pattern to distribute the bindings more easily.

BUG=
Change-Id: I96ff7c3715164c533202ce300ab0312b1b07f884

Change-Id: I96ff7c3715164c533202ce300ab0312b1b07f884
Reviewed-on: https://skia-review.googlesource.com/10032
Reviewed-by: Mike Klein <mtklein@chromium.org>
Commit-Queue: Stephan Altmueller <stephana@google.com>

Former-commit-id: efa48d599de15d960d7335a4a93a6bbeb97d3c41