[Offline pages] Stop observing web contents and reset to null after use. Only create a new web contents when we need it. (M59 merge)
Before this change, we will create a new webcontents in preparation for a request that might not come in for hours or days. During the waiting period, a number of things can happen which may result in things related to the webcontents to become stale and therefore crash when we actually do use it. One hypothesis is that a render process host might be created without any render processes attached to it, causing the host to be removed under certain memory pressures or when certain clean-up processes are triggered.

For better memory and freshness, we should probably only create a web contents when we actually need it anyway, and clean up everything after we use it.

BUG=714272,709443

Review-Url: https://codereview.chromium.org/2866643005
Cr-Original-Commit-Position: refs/heads/master@{#469906}
Review-Url: https://codereview.chromium.org/2873393004 .
Cr-Commit-Position: refs/branch-heads/3071@{#521}
Cr-Branched-From: 91bcc2d22c7c7ff519c4f50384d9bdc5a08671e0 [formerly a106f0abbf69dad349d4aaf4bcc4f5d376dd2377]-refs/heads/master@{#464641}

Former-commit-id: 339b80c4f9ff8e29b73b0842570be8b852f2340f