Add check for incorrect opList splits
This CL relies on: https://skia-review.googlesource.com/c/14186/ (Split up opLists (take 3))

We can't/don't want to close the opList in the GrSurfaceContext dtor because we, semi-often, wrap a proxy in a utility SurfaceContext in order to upload data to it (e.g., when creating a proxy w/ data or in the texture strip atlas).

In the current set up, the "temporary fixes" in newRTOpList and newTextureOpList should prevent any out of order execution.

Change-Id: I8189db6ed59096d7b8c1704028df395f84e4a681
Reviewed-on: https://skia-review.googlesource.com/14187
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>

Former-commit-id: a4c93ac535d3b0ed979c7bf666f09aa7dbf1f0eb