comment SK_API and add fix params
I thought that SK_API on operator== was 
redundant with SK_API on the enclosing class,
and was not needed. Turns out for mac_chromium_debug_ng
it is needed. Added comments for history.

Meanwhile, bookmaker found some missing and
mis-named parameters, so all is not lost.

R=reed@google.com

Change-Id: I88645666a9d06ec90c5ac133673460d6e6c75528
Reviewed-on: https://skia-review.googlesource.com/17277
Commit-Queue: Cary Clark <caryclark@google.com>
Reviewed-by: Mike Reed <reed@google.com>

Former-commit-id: 0e616cf9e34b306d53fb69a4db76fee02fd1dcc3