streamline SkRasterPipeline::run()
This used to need to be complicated to check for unimplemented stages,
but now that SkJumper rules the world, we can make this a lot simpler
and move a lot of the tricky stage selection logic into an SkOnce.

This should decrease the overhead of SkRasterPipeline::run().

Change-Id: I79d73f3315c7a1a52ab02747a5f17ce27d8cb9a4
Reviewed-on: https://skia-review.googlesource.com/17440
Reviewed-by: Herb Derby <herb@google.com>
Commit-Queue: Mike Klein <mtklein@chromium.org>

Former-commit-id: 9f52e98e60ea763c5d6f80e7d7fbc6da7c48559b