Detect frames from rotated devices in VideoTrackAdapter on Android.
On Android, when a device is rotated, frames from the camera arrive with
width and height swapped. The VideoTrackAdapter was unable to detect this
and instead cropped them incorrectly because it treated the rotated
dimension as a violation of the maximum size settings.

This CL provides a simple mechanism to detect this problem on Android.
If the frame arrives from the camera with width and height swapped from
what was expected, treat it as a valid frame that does not need to be
adjusted.

BUG=718823

Review-Url: https://codereview.chromium.org/2870413004
Cr-Original-Commit-Position: refs/heads/master@{#472057}
Review-Url: https://codereview.chromium.org/2897993002 .
Cr-Commit-Position: refs/branch-heads/3071@{#652}
Cr-Branched-From: 91bcc2d22c7c7ff519c4f50384d9bdc5a08671e0 [formerly a106f0abbf69dad349d4aaf4bcc4f5d376dd2377]-refs/heads/master@{#464641}

Former-commit-id: a110813cbb7d0f8ff2395a6a4138183e97946ea5