Fix Docs links from Files app on ChromeOS not working with PlzNavigate.
The cause was that there was a redirect from externalfile scheme to https. The PlzNavigate code chose the right ending process based on the final URL. However when the renderer made a request for the stream URL, it does attach the original URL (pre-redirect) along. ResourceDispatcherHostImpl::ShouldServiceRequest was incorrectly aborting the URL because that process rightly does not have access to the externalfile scheme.

The fix is to skip the ChildProcessSecurityPolicy::CanRequestURL check in ShouldServiceRequest for the PlzNavigate stream request. The browser already picked the right process, and the stream URL is unguessable. The change also skips most of ContinuePendingBeginRequest() and CreateResourceHandler() for PlzNavigate stream requests to minimize the risk that they contain code that could be exploited.

BUG=717644

Review-Url: https://codereview.chromium.org/2908593002
Cr-Original-Commit-Position: refs/heads/master@{#475097}
Review-Url: https://codereview.chromium.org/2910063003 .
Cr-Commit-Position: refs/branch-heads/3112@{#31}
Cr-Branched-From: 07faed0551ea9474531c25080552dfc7bccfa9e8 [formerly b6460e24cf59f429d69de255538d0fc7a425ccf9]-refs/heads/master@{#474897}

Former-commit-id: 12c73ea1a9ada42efdc3e2b245e718214d6f3125