assume sRGB transfer function on untagged images
This has been a long time coming.  There are probably one or two more
places we'd want to do this on the CPU side too, but it's a start.

Somewhat seriously, this is my proposed fixed for colorburn +
xfermodes2.  We're not interpreting the red or green images sensibly.

BUG=skia:6680

Change-Id: Iea15be08c1907ac1ac46ef358d3d74f1cc35a156
Reviewed-on: https://skia-review.googlesource.com/18221
Reviewed-by: Brian Osman <brianosman@google.com>
Reviewed-by: Matt Sarett <msarett@google.com>
Commit-Queue: Mike Klein <mtklein@chromium.org>

Former-commit-id: 9c6645e787669e5af2b9de73d0a6d1cd4d77bbf0