Add username field discovery heuristic.
There are some cases when autofill can't find a username field for
autocompletion. For example, if the field is created dynamically in JavaScript.
This CL introduces a heuristic which is used as a fallback in such cases to
find a username field.

When password form autocompletion is about to happen with unknown username
field, the heuristic searches for the closest visible autocompletable
non-password text element preceding the password element. If one found, it is
believed to be the username field.

BUG=710374,606736,586465,710003

Review-Url: https://codereview.chromium.org/2889393002
Cr-Original-Commit-Position: refs/heads/master@{#475481}
Review-Url: https://codereview.chromium.org/2918543002 .
Cr-Commit-Position: refs/branch-heads/3112@{#52}
Cr-Branched-From: 07faed0551ea9474531c25080552dfc7bccfa9e8 [formerly b6460e24cf59f429d69de255538d0fc7a425ccf9]-refs/heads/master@{#474897}

Former-commit-id: 525e06d30faf3fdbfb85e2b0e1e7c0c1bdc031e5