VR: Content size incorrect when transitioning from webVR into VR Shell
The issue here is we are setting the virtual display's size when we
enter WebVR, but not when we exit it to VrShell.  This means that
content is incorrectly sized.

The fix simply sets the virtual display's size when we exit WebVR to VrShell.

BUG=718578
NOTRY=true
NOPRESUBMIT=true

Review-Url: https://codereview.chromium.org/2906173002
Cr-Original-Commit-Position: refs/heads/master@{#476074}
Review-Url: https://codereview.chromium.org/2921583003
Cr-Commit-Position: refs/branch-heads/3112@{#108}
Cr-Branched-From: 07faed0551ea9474531c25080552dfc7bccfa9e8 [formerly b6460e24cf59f429d69de255538d0fc7a425ccf9]-refs/heads/master@{#474897}

Former-commit-id: 079395cb4a1ae3e6267aada372111300602ceb2e