chrome/android: Fix black flash on cold start.
We currently see a black flash on cold start with some pages because
the logic uses the SwapBuffers notification to hide the background color
on the View to make the underlying Surface visible.

Since the SwapBuffers notification doesn't ensure that the rendering
has finished, it may not actually be used by the SurfaceFlinger which
causes a black flash.

Since we have similar logic for the tab strip which is also drawn in
native, merge this logic in one place and remove the background only
after the second frame is drawn.

BUG=729278

Review-Url: https://codereview.chromium.org/2926803003
Cr-Original-Commit-Position: refs/heads/master@{#477774}
Review-Url: https://codereview.chromium.org/2928703006 .
Cr-Commit-Position: refs/branch-heads/3112@{#244}
Cr-Branched-From: 07faed0551ea9474531c25080552dfc7bccfa9e8 [formerly b6460e24cf59f429d69de255538d0fc7a425ccf9]-refs/heads/master@{#474897}

Former-commit-id: d317f0434ffd4eacbe8f4f412da254a411cf34ea