Fix 'Less than 1 MB' used inside a sentence
The string 'Less than 1 MB' is used inside a sentence in CBD.
This is incorrect because of the upper case L.
To fix this we reintroduce the "less than 1 MB" string. It is still in
the translation console, so it shouldn't require translation.

This CL doesn't fix the usage of 'Less than X MB' which is only used
on desktop and the fact that this sentence insertion could go wrong
due to  sentence structure in other languages.

BUG=729942

Review-Url: https://codereview.chromium.org/2921423002
Cr-Original-Commit-Position: refs/heads/master@{#477267}
Review-Url: https://codereview.chromium.org/2928933002 .
Cr-Commit-Position: refs/branch-heads/3112@{#248}
Cr-Branched-From: 07faed0551ea9474531c25080552dfc7bccfa9e8 [formerly b6460e24cf59f429d69de255538d0fc7a425ccf9]-refs/heads/master@{#474897}

Former-commit-id: 10a2061f67a6ca24c511ed05339feae8bf57f6d5