Cancel dialogs for Tabs when they are removed from their TabModels.
A WebState attempts to cancel its dialogs in its destructor.  However,
the BVC uninstalls itself as the WebStateDelegate before the WebState
is deallocated, so the WebStateImpl doesn't have access to its
JavaScriptDialogPresenter to handle the cancellation.

This CL adds a call to |-cancelDialogForWebState:| when the delegates
are uninstalled for a Tab to ensure that alerts aren't queued in the
DialogPresenter for WebStates that have been deallocated.

BUG=720778

Review-Url: https://codereview.chromium.org/2926413004
Cr-Original-Commit-Position: refs/heads/master@{#478729}
Review-Url: https://codereview.chromium.org/2934003003 .
Cr-Commit-Position: refs/branch-heads/3112@{#328}
Cr-Branched-From: 07faed0551ea9474531c25080552dfc7bccfa9e8 [formerly b6460e24cf59f429d69de255538d0fc7a425ccf9]-refs/heads/master@{#474897}

Former-commit-id: 7655bb4c57d3ecdbaf03a1c46453b83e879a04f4