Merge to M60: Add null check when reporting changed nodes to an AX tree.
See bug for repro; this happens when we get an event fired on a window
that's not part of the window hierarchy. The rest of the code ignores it
fine but the tree change callbacks get called on a null node, which
causes problems.

BUG=727842

Review-Url: https://codereview.chromium.org/2929673002
Cr-Original-Commit-Position: refs/heads/master@{#482381}
Review-Url: https://codereview.chromium.org/2971963004 .
Cr-Commit-Position: refs/branch-heads/3112@{#532}
Cr-Branched-From: 07faed0551ea9474531c25080552dfc7bccfa9e8 [formerly b6460e24cf59f429d69de255538d0fc7a425ccf9]-refs/heads/master@{#474897}

Former-commit-id: 95970388a9b75f34d1a86f2dec8ebebb6a7d084b