Ensure FB sRGB conversion is on during copy-as-draw
A recent change up-stack tickled the cross context image test, exposing
this bug. I have no idea how it didn't manifest sooner. In that test,
we're takng a subset of a lazy image, which implements onGenerateTexture
via a surface copy. In this case, we were in legacy mode, but the
textures were all created with sRGB configs. In normal drawing, we
disable sRGB decode. For the copy draw, we were explicitly enabling sRGB
decode, but not setting the FB setting, so if it was still off (it was),
the image would get decoded but not re-encoded, coming out too dark.

Bug: skia:
Change-Id: I4a480b11315cc308ea016b7082aff17920764214
Reviewed-on: https://skia-review.googlesource.com/22211
Commit-Queue: Brian Osman <brianosman@google.com>
Reviewed-by: Brian Salomon <bsalomon@google.com>

Former-commit-id: f02fa6ffe72c596b87e277193ed82d288dbee18f