when encoding PNG, don't assume F16 values are in range
I believe this will fix this batch of broken colorImages:

   https://gold.skia.org/search?blame=671a63676178747b5f094fd40c1fbcf528cee346 [formerly e23e55ef33358b2f6f98fc9cb795c71397c01618]&fdiffmax=-1&fref=false&frgbamax=255&frgbamin=0&head=true&include=false&limit=50&match=gamma_correct&match=name&metric=combined&neg=false&offset=0&pos=false&query=source_type%3DcolorImage&sort=desc&unt=true

Depending on the instruction set, clamping to logical [0,1] may be
implicit or not as we work our way down to bytes.  I think that's why
these broken images are not more widespread on the bots.

Change-Id: Ie9bb937864bf6954301d76d9921a2d3029d12c9a
Reviewed-on: https://skia-review.googlesource.com/25742
Reviewed-by: Leon Scroggins <scroggo@google.com>
Commit-Queue: Mike Klein <mtklein@chromium.org>

Former-commit-id: ce98801bbb5ac251b374f8b565445add43017aaa