_very_ minor srcover speedup
This is a consistent, very small speedup for srcover.
    SkRasterPipeline_run
    Before: 30.4057ns
    After:  30.1089ns
i.e. a 1% speedup on the bench, maybe 3-4% improvment in srcover itself.

The only reason I'd send this out now is that this will slightly change
some pixels, so it's a good thing to sneak in before rebaselining.

It's possible that other blend modes would benefit from the same, but
I've only looked at srcover (and I've also changed dstover so that it
doesn't look funny).

Change-Id: Ic056ca0912d76648d43a78e0052176fd0f7934f1
Reviewed-on: https://skia-review.googlesource.com/30281
Reviewed-by: Florin Malita <fmalita@chromium.org>
Commit-Queue: Mike Klein <mtklein@chromium.org>

Former-commit-id: c18ab884ae6cf85521ed2c719d78e45308404335