Don't validate() in ~SkPathRef
If an SkPathRef is constructed from arbitrary memory, it might not be
valid.  However, any SkPathRef should be able to be destroyed cleanly.

An alternative to this would be to have CreateFromBuffer always do
the validation (even in release builds), return null for invalid ones,
and do the cleanup itself, but this seems like extra complication and
maybe not something general release builds want.

Bug: chromium: 752755
Change-Id: I1d509a5d5d0b173c20162ff79d731d6c8b4b6fb4
Reviewed-on: https://skia-review.googlesource.com/37321
Commit-Queue: Adrienne Walker <enne@chromium.org>
Commit-Queue: Mike Klein <mtklein@google.com>
Reviewed-by: Mike Klein <mtklein@google.com>

Former-commit-id: d6b28f7bc8c84ee952c433d54861620e403b990f