Fix a regression on Windows 7 which manifested in normal windows obscuring the taskbar when a fullscreen window is around.
Merging to M61

This bug only occurs on Windows 7 and started with my change where we use the ITaskbarList2::MarkFullscreenWindow() API
to inform the taskbar about a window being fullsceen or not. As per docs calling this API when a window goes fullscreen is enough
for the taskbar to hide itself when that window is active and to show itself when the window is inactive. It appears that the behavior
is different on Windows 7.

The fix being tried here is to mark the window as fullscreen when it is active and inactive when it is not, i.e a manual process.
This works fine on Windows 7 and 10 and was verified by QA.

BUG=758113
TBR=ananta@chromium.org

(cherry picked from commit 82689a54a747725acdbebee9d12a56e523c02ba5 [formerly f091c0c1b2a968a5f61dc15da2b35b87077d077a])

Change-Id: Ibdeae0be10640559594fe97980b622d64ae6ae75
Reviewed-on: https://chromium-review.googlesource.com/634451
Commit-Queue: Ananta Iyengar <ananta@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#497593}
Reviewed-on: https://chromium-review.googlesource.com/640070
Reviewed-by: Ananta Iyengar <ananta@chromium.org>
Cr-Commit-Position: refs/branch-heads/3163@{#961}
Cr-Branched-From: e88a7fd797602ff08086e1663443abb227a71a68 [formerly ff259bab28b35d242e10186cd63af7ed404fae0d]-refs/heads/master@{#488528}
Former-commit-id: 854c089d7063043bc260865aeeb77dc6a1d1a793