The device policy functionality for ecryptfs to ext4 migration removed
Starting from version 61, the user policy controls if the migration
from ecryptfs to ext4 is allowed. The device policy that was
managing this functionality in version 60 now has to be removed.

The ARC icon is displayed based on IsArcAllowedForProfile function. It
was updated to return false if the encryption is ecryptfs and the
migration is forbidden for the user.

The code for removing ARC data, which was forbidden in previous
version if the device policy was disallowing the migration is
removed now. In case we decide to make another exception there, it will
be implemented in a separate CL.

Also the banner stating that the migration is required now will be
displayed if the partition is ecryptfs. The logic related to the
banner is to be decided and implemented in a separate CL.

TBR=igorcov@chromium.org

(cherry picked from commit f1f81f356906f981aabf08607a20646dbc602155 [formerly 1d6a3bf3e05ee228fb27ae0f1d91edd91f24cabe])

Bug: chromium:725493
Test: Unit tests and manual test.
Change-Id: Id5a10c875d2c1728169df01fd0fca095595fce6e
Reviewed-on: https://chromium-review.googlesource.com/582951
Commit-Queue: Igor <igorcov@chromium.org>
Reviewed-by: Paweł Hajdan Jr. <phajdan.jr@chromium.org>
Reviewed-by: Igor <igorcov@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Kazuhiro Inaba <kinaba@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#498448}
Reviewed-on: https://chromium-review.googlesource.com/650249
Cr-Commit-Position: refs/branch-heads/3163@{#1097}
Cr-Branched-From: e88a7fd797602ff08086e1663443abb227a71a68 [formerly ff259bab28b35d242e10186cd63af7ed404fae0d]-refs/heads/master@{#488528}
Former-commit-id: 86ea32bd3e120afb5ad9bd911750542cf3b36ef5