NTP on iPad can be scrolled up to the fake omnibox
The height of the ContentSuggestions collection on NTP iPad is not
sufficient to allow the fake omnibox to be scrolled enough to be
hidden by the real toolbar.
This is particularly a problem when the omnibox is focused because
the collection is only half scrolled if the content is not high enough.

Bug: 766483
Change-Id: I00c2b6f7446e801dec56e19af28d139da42b8b1a
Reviewed-on: https://chromium-review.googlesource.com/671354
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Jean-François Geyelin <jif@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#503080}(cherry picked from commit ef047166e366b5697b78fd902ecba04bbbd3fd96 [formerly 1a833e0ec4edb771ff87c4cd98625eb8b1e1d9bf])
Reviewed-on: https://chromium-review.googlesource.com/678674
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/branch-heads/3202@{#393}
Cr-Branched-From: 3efa1198dfb9f65f00675896e8ea5a4db4f46842 [formerly fa6a5d87adff761bc16afc5498c3f5944c1daa68]-refs/heads/master@{#499098}
Former-commit-id: 8b06611aaf15bbf5fe19f99cb2e3c394fec3c380