Also mask t itself in mask_2pt_conical_degenerates.
We look at t to create a mask in mask_2pt_conical_degenerates to be
applied later to the colors after the normal gradient stages have run.

But if t itself is NaN, that will wreak havoc in the normal gradient
stages.  So in addition to building the mask to kill off degenerate
colors, let's also set degenerate t to zero, which should be a safe
value.

This fixes the fuzz mentioned in this bug.
BUG=skia:7078

Change-Id: I8301450c707bdbf941abd0339959f9e60d46d955
Reviewed-on: https://skia-review.googlesource.com/52763
Reviewed-by: Florin Malita <fmalita@chromium.org>
Commit-Queue: Mike Klein <mtklein@chromium.org>

Former-commit-id: 8477144e75329b146feb2a14cc04696aa53e0a90