When flushing, reset flush state before op lists
When we reset the flush state after executing the ops in a flush, we reset
the stored pipelines which may be holding refs to resources in their processors.

We want the resources that we wrote to in the op list to be freed last for caching
LRU purposes. Thus we make sure to free all the pipeline resources first before
resetting and freeing the op lists

Bug: skia:
Change-Id: I5b02e9b7532007b5854e906bf4d64e901e7e7e53
Reviewed-on: https://skia-review.googlesource.com/54760
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Greg Daniel <egdaniel@google.com>

Former-commit-id: 78bdee200c26520eb065728ae7c094b76ecaae66