color: Make transfer function approximation continuous
When approximating the transfer function's linear segment, ensure that
- it go through the origin
- it be continuous
This is in contrast with the previous behavior, which was to do a linear
fit.

This may result in a greater error than otherwise (e.g, if the table's
data didn't go through the origin). That does not matter, because we
will then judge the fit by its maximum error, and if the maximum error
is too large, we will use a LUT at compositing time.

As an opportunistic fix, when doing the nonlinear least squares on the
nonlinear segment of the data, assume that all in the domain [0.15, 1.0]
be in the nonlinear segment. This is in contrast to the previous
behavior, which used the domain [0.25, 1.0]. I found repeatedly that
extending this domain resulted in a better fit (and the linear segment
starts almost-never-if-not-never above 0.15).

TBR=ccameron@chromium.org

(cherry picked from commit c7c8022ab40891458291084991441975b15c4198 [formerly c3dfdcdcba1f5fbd1f95e5e1fa0c6396ad58fbdc])

Bug: 768724
Change-Id: Id76cf380c08bc8b050bbb4fd7c9668a186fa447e
Reviewed-on: https://chromium-review.googlesource.com/695181
Reviewed-by: Fredrik Hubinette <hubbe@chromium.org>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#506144}
Reviewed-on: https://chromium-review.googlesource.com/703454
Reviewed-by: ccameron chromium <ccameron@chromium.org>
Cr-Commit-Position: refs/branch-heads/3202@{#598}
Cr-Branched-From: 3efa1198dfb9f65f00675896e8ea5a4db4f46842 [formerly fa6a5d87adff761bc16afc5498c3f5944c1daa68]-refs/heads/master@{#499098}
Former-commit-id: f979f378ef8a8aae6500382ec256b78f156e7279