Convert MediaTrackConstraints to a ScriptValue
IDLDictionaries such as MediaTrackConstraints should not be stored on
the heap which would happen when binding one as a parameter to a
callback. This change converts the object to a ScriptValue ahead of
time. This is fine because the value will be passed to a
ScriptPromiseResolver that will converted it to a V8 value if it
isn't already.

Bug: 759457
Change-Id: I3009a0f7711cc264aeaae07a36c18a6db8c915c8
Reviewed-on: https://chromium-review.googlesource.com/701358
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#507177}(cherry picked from commit 4bcb667da30d56c5d59d3f192c1e929b70367f1c [formerly 84ca1ee18bbc32f3cb035d071e8271e064dfd4d7])
Reviewed-on: https://chromium-review.googlesource.com/710173
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/branch-heads/3202@{#642}
Cr-Branched-From: 3efa1198dfb9f65f00675896e8ea5a4db4f46842 [formerly fa6a5d87adff761bc16afc5498c3f5944c1daa68]-refs/heads/master@{#499098}
Former-commit-id: 19765bd888949485b1e7ee2892b6502b1a653025