Disable fading text in omnibox on iOS 11.1.
A bug in iOS 11 broke the omnibox most significant part of the URL
display mechanism; and in response to the radar in 11.1 the "fix" is
changing the way drawTextInRect: operates - now it just ignores its
argument.
This causes issues with omnibox that is overriding this method in order
to mask out the clipped prefix/suffix with a fade effect and to display
the most significant part of the URL: the URL now cannot be cut the way
we want, so the fade effect obscures useful text in the omnibox instead
This CL just disables this custom behavior on 11.1.

Bug: 771248
Change-Id: I9628adf195c3cbec36b962d2f872f658ad960529
Reviewed-on: https://chromium-review.googlesource.com/704638
Reviewed-by: Rohit Rao (ping after 24h) <rohitrao@chromium.org>
Commit-Queue: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#507046}(cherry picked from commit bd9e349b27f8ad70fa6c789379d36df571a918b4 [formerly 386b9b2dad021739ef321fa485a54238586bc658])
Reviewed-on: https://chromium-review.googlesource.com/718357
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Commit-Position: refs/branch-heads/3202@{#679}
Cr-Branched-From: 3efa1198dfb9f65f00675896e8ea5a4db4f46842 [formerly fa6a5d87adff761bc16afc5498c3f5944c1daa68]-refs/heads/master@{#499098}
Former-commit-id: ecec8cdd5559b8a6ad77a7de13732a4c9ee9e39c