Remove subpixel positioning implies no bytecode hinting.
SkTypeface_FreeType::onFilterRec currently assumes that if asked to do
subpixel positioning, don't do bytecode hinting. The idea was that both
could not be satisfied at the same time, so pick something. This is no
longer true, as with the v40 interpreter it is possible to get subpixel
positioned but bytcode hinted glyphs.

BUG=skia:6931

Change-Id: I4aef1670abc75623ad29bab0716d5ec9ebd2d906
Reviewed-on: https://skia-review.googlesource.com/59984
Reviewed-by: Ben Wagner <bungeman@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>

Former-commit-id: 2d6b3937e04dcd1ab7f1e9a7f49f9bb3fd43b261