cc: Only populate image to frame index map with cc image animations.
Currently we select the same frame index to use across all PaintImages
on a tile, which comes from blink on the PaintImage when image
animations are driven by blink. If this tile is partially offscreen
such that one image instance is offscreen, then blink never updates
the recording for it. So depending on the order in which we visit
images on the tile, we may select the frame index for the offscreen
instance of the image, which will effectively pause the animation also
for the visible instance of the image on this tile.

In order to avoid this, only populate the map when compositor image
animations are enabled. This can still lead to cases where if cc is
selecting the frame index to use, we attempt to use an index with an
instance of PaintImage that does not have data for that frame, which
will result in a crash in the generator when we query the FrameKey for
this frame. But the correct fix for this needs to be in blink, to
update the recording when the data for the image changes and ensure
all instances of PaintImage received in a commit share the same
generator.

R=vmpstr@chromium.org
TBR=khushalsagar@chromium.org

(cherry picked from commit f43ad04817d39f329233a1b1b87c5f5466203649 [formerly 58237fb768d3c6951396f6d162958526847f7924])

Bug: 768975, 775558
Cq-Include-Trybots: master.tryserver.blink:linux_trusty_blink_rel
Change-Id: Iae302875166d667172018d0835370792bbd7a27c
Reviewed-on: https://chromium-review.googlesource.com/726742
Commit-Queue: Khushal <khushalsagar@chromium.org>
Reviewed-by: vmpstr <vmpstr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#510388}
Reviewed-on: https://chromium-review.googlesource.com/732734
Reviewed-by: Khushal <khushalsagar@chromium.org>
Cr-Commit-Position: refs/branch-heads/3239@{#144}
Cr-Branched-From: 393459ba3d4b7de1340326f1b415a9c7ebf06d39 [formerly adb61db19020ed8ecee5e91b1a0ea4c924ae2988]-refs/heads/master@{#508578}
Former-commit-id: 618173b110e090930ea879180fca035d5e5f5a85