Android NTP: don't "throw IllegalStateException()"
The ntp_tiles component is not supposed to return multiple tiles with
the same URL. In order to determine if that still happened in the wild,
the throw statement was added, with the intention of catching the
problem in Beta. However, it seems that Beta hasn't had trouble with
this, but now it's a major crasher on Stable.

It looks to me that it's sufficient to simply remove the throw and
replace it with a continue.

Bug: 703628
Change-Id: If5d1b24a8e69b34d92ae6558c7bcbffad5995906
Reviewed-on: https://chromium-review.googlesource.com/734040
Commit-Queue: Chris Pickel <sfiera@chromium.org>
Reviewed-by: Nicolas Dossou-Gbété <dgn@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#511086}(cherry picked from commit 47e44e6cd336193033f25620873b0df295c35174 [formerly 734fa1a007546b823df0e539ca1340e27156a428])
Reviewed-on: https://chromium-review.googlesource.com/735612
Reviewed-by: Chris Pickel <sfiera@chromium.org>
Cr-Commit-Position: refs/branch-heads/3239@{#177}
Cr-Branched-From: 393459ba3d4b7de1340326f1b415a9c7ebf06d39 [formerly adb61db19020ed8ecee5e91b1a0ea4c924ae2988]-refs/heads/master@{#508578}
Former-commit-id: c716bc2910e88e25752545ea1fbfe0a68a7b66e3