Remove Nexus10 GPU jobs.
The latest OS is Lollipop, and it was originally released November 2012.
The GPU is not in the top 26 according to UMA stats (I haven't collected
data for lower than 26). We test the same GPU with the
Chromebook_303C12, and bsalomon says "Personally, I'm OK with just
relying on ChromeOS for testing."

No-Try: true
Change-Id: Ib17033153faab9d99613e7d2d069524c9435f65e
Reviewed-on: https://skia-review.googlesource.com/64066
Reviewed-by: Brian Salomon <bsalomon@google.com>
Commit-Queue: Ben Wagner <benjaminwagner@google.com>

Former-commit-id: 1a2b3f0f93b40d4451616905d821f6cd3e1373a8