Directly use SkScalerContextRec.
Some time ago SkScalerContext::Rec was made SkScalerContextRec so that
it could be forward declared. However, SkScalerContext::Rec remains as a
typedef to SkScalerContextRec. This removes the typedef and updates the
users to use the SkScalerContextRec type directly.

This change was prompted by my own confusion when looking at some code
which mixed these two, by IntelliSense's confusion when declaring an
argument with one and defining with the other, and reducing the general
proliferation of nested 'Rec' types in Skia.

Change-Id: I3a23a4bdd83a591807c78cea6aa7a0117544b614
Reviewed-on: https://skia-review.googlesource.com/66153
Reviewed-by: Herb Derby <herb@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>

Former-commit-id: 3e45a2f48ed8dbf3a55e980c43d97d29afeb93a0