Enforce gradient color stop monotonicity in ctor
From the SVG 1.1 spec:
Each gradient offset value is required to be equal to or greater than the previous 
gradient stop's offset value. If a given gradient stop's offset value is not
equal to or greater than all previous offset values, then the offset
value is adjusted to be equal to the largest of all previous offset values.

Change-Id: I797369a1e14dc776ceb6478ac9fcdd4792e65562
Reviewed-on: https://skia-review.googlesource.com/67761
Commit-Queue: Florin Malita <fmalita@chromium.org>
Reviewed-by: Herb Derby <herb@google.com>
Reviewed-by: Mike Klein <mtklein@chromium.org>

Former-commit-id: 3e20d023bfa6e534ee0abacdba225dd140ae471c