Fix toolbar leading calculation on RTL
The toolbar has two view: the background view and the content view.
The content view has the same size as the background view using auto
layout. When the frame of the toolbar is directly modified, the
constraints do not kick in before we use the size of the background
view to resize the omnibox.
This creates a problem as the content inside the content view is using
the autoresizing mask. When the content view is resized with auto
layout, its content is resized a second time.
This CL fixes it by using the frame of the content view instead of the
frame of the background view.

Bug: 781929, 781910
Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet
Change-Id: Ia42190363ad48d08422b085e55a4a94e216ec9a9
Reviewed-on: https://chromium-review.googlesource.com/756708
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#514440}(cherry picked from commit caee47fe274583257358c4df341369717be00639 [formerly fd9517b66d8ed921e16d3c3ebe64c8fd45eb7ca0])
Reviewed-on: https://chromium-review.googlesource.com/763087
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/branch-heads/3239@{#439}
Cr-Branched-From: 393459ba3d4b7de1340326f1b415a9c7ebf06d39 [formerly adb61db19020ed8ecee5e91b1a0ea4c924ae2988]-refs/heads/master@{#508578}
Former-commit-id: 1a69914ad11beb5b6f6dd2e111114caaa5018033