Eliminate the use of a tmp buffer for image blur
Because the blur_one_direction routine, only reads the
source buffer once for each result, it can be used to
blur inplace. This allows the code to avoid using
a temporary buffer for the intermediate result.

This produces the same pixels according to skdiff, and
has the same speed.

Change-Id: I6485a71f0eb39936d69b8a98ab429a2cf592a7cd
Reviewed-on: https://skia-review.googlesource.com/74520
Reviewed-by: Florin Malita <fmalita@chromium.org>
Commit-Queue: Herb Derby <herb@google.com>

Former-commit-id: a0ec54829eb77ec51747d2c8fd1ecfb7c0587965