Enable DAA for MSVC
It seems that we don't have to worry about MSVC's performance.
If we're worried, we can revert this CL after the run of our perf
bots so I can get the latest performance benchmark.

A large set of golden images in MSVC config are expected to change
with this CL.

TBR: mtklein@google.com

Merge branch 'master' of https://skia.googlesource.com/skia


Change add to toggle in the comments

Bug: skia:
Change-Id: If682673bab558f87196d4437c405732ba2ffcf87
Reviewed-on: https://skia-review.googlesource.com/74400
Reviewed-by: Yuqian Li <liyuqian@google.com>
Commit-Queue: Yuqian Li <liyuqian@google.com>

Former-commit-id: 61170a07191f0483ff405afc620cb0d554063375