Fix pause then cancel download issue for API < 21.
There was a bug that it was not possible to cancel a notification after
the download was paused. This is because in API < 21, calling
stopForeground(false) would make the notification continue to be
classified as on-going and therefore undismissable, which is what was
happening in the pause case. This CL fixes this problem by inducing a
flicker that dissociates the notification from the foreground service
in the case that the Android version is API < 21.

Bug: 788080
Change-Id: I9c8776d245425b0fe78aaccf7e75fcfb6e57f039
Reviewed-on: https://chromium-review.googlesource.com/802078
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Joy Ming <jming@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#521043}(cherry picked from commit 074e6b5f2136eb99a491d059c36fe381e35d5737 [formerly baefcf8a84ae1f6e81f31f14319f00f0dac839ae])
Reviewed-on: https://chromium-review.googlesource.com/809305
Reviewed-by: Min Qin <qinmin@chromium.org>
Cr-Commit-Position: refs/branch-heads/3282@{#34}
Cr-Branched-From: ab27e3b6feefb91d7474fe53634cf832758e747e [formerly 5fdc0fab22ce7efd32532ee989b223fa12f8171e]-refs/heads/master@{#520840}
Former-commit-id: 8dabd604660419d64ea279412f224b429d9f3c8b