Fix Android WebView crash when an app loads a JavaScript URL during navigation.
When there's an ongoing navigation, if the app loada a JavaScript URL we create
a temporary NavigationRequest in NavigatorImpl::RequestNavigation. However the
call to RenderFrameHostManager::GetFrameHostForNavigation was resetting the
speculative RFH of the original NavigationRequest since the site instance of a
JavaScript load is always the same as the existing frame.

This is a reland of r523264 with an extra fix to ensure that the RFH is
initialized.

Bug: 793432
Change-Id: I7a029d907f5c133c00ce096d9d733cb8194fcd43
Reviewed-on: https://chromium-review.googlesource.com/822970
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Commit-Queue: John Abd-El-Malek <jam@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#523489}(cherry picked from commit c65502707c3b25efc2d2618b7da13b5229f98230 [formerly 4aff0098fa34105dc76d51c32162c40c34490391])
Reviewed-on: https://chromium-review.googlesource.com/823330
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Cr-Commit-Position: refs/branch-heads/3239@{#665}
Cr-Branched-From: 393459ba3d4b7de1340326f1b415a9c7ebf06d39 [formerly adb61db19020ed8ecee5e91b1a0ea4c924ae2988]-refs/heads/master@{#508578}
Former-commit-id: 110a05ed12c7c8f1bcb686e0c01e37c9d23b8922