Fix Android WebView crash when an app loads a JavaScript URL during navigation.
When there's an ongoing navigation, if the app loada a JavaScript URL we create
a temporary NavigationRequest in NavigatorImpl::RequestNavigation. However the
call to RenderFrameHostManager::GetFrameHostForNavigation was resetting the
speculative RFH of the original NavigationRequest since the site instance of a
JavaScript load is always the same as the existing frame.

This is a reland of r523264 with an extra fix to ensure that the RFH is
initialized.

Bug: 793432
Change-Id: I7a029d907f5c133c00ce096d9d733cb8194fcd43
Reviewed-on: https://chromium-review.googlesource.com/822970
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Commit-Queue: John Abd-El-Malek <jam@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#523489}(cherry picked from commit c65502707c3b25efc2d2618b7da13b5229f98230 [formerly 4aff0098fa34105dc76d51c32162c40c34490391])
Reviewed-on: https://chromium-review.googlesource.com/823350
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Cr-Commit-Position: refs/branch-heads/3282@{#189}
Cr-Branched-From: ab27e3b6feefb91d7474fe53634cf832758e747e [formerly 5fdc0fab22ce7efd32532ee989b223fa12f8171e]-refs/heads/master@{#520840}
Former-commit-id: 277e60c0c0a41d56f0d9312bbb0b71b9c98a845b