fix very large clipped path limit
Mozilla notes that clipped paths conservatively triple
the reserved space for a path edge list, potentially
overflowing an int if the point count is 2^31/3 or
larger, making maxEdgeCount negative if maxEdgeCount
is an int.

By making maxEdgeCount size_t, the multiply stays in
range. A couple of lines down, makeArrayDefault is going
to trigger an SkASSERT_RELEASE because the record size
times the point count exceeds the allowable limit.

R=scroggo@google.com
Bug: skia:7391
Change-Id: Ib20b392a369133c91fe2785be248dce3a2100202
Reviewed-on: https://skia-review.googlesource.com/85720
Commit-Queue: Cary Clark <caryclark@google.com>
Reviewed-by: Leon Scroggins <scroggo@google.com>

Former-commit-id: 29c14a760682e2c449fa043b5e8b69937cb58f3a