Remove conversion to 8888 of sRGB sources with other configs
Our color space xform code now handles applying the transfer function in
the shader if there's a mismatch between the pixel config and the color
space, so this logic is no longer needed. Although it may produce
slightly less accurate filtering, the tradeoff of (for example)
expanding grayscale bitmaps by a factor of four, on the off chance that
we'll draw them to a color managed destination isn't worth it.

Bug: skia:
Change-Id: Ic56e7d48301cd7cbb0d5bccd17cfd383786f98e5
Reviewed-on: https://skia-review.googlesource.com/90901
Reviewed-by: Brian Salomon <bsalomon@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>

Former-commit-id: 9d44081e00f67361414935b82332df5754bddcaf