[Merge M64] Extensions: Only register live RenderFrameHosts with the ProcessManager.
This is a speculative fix for issue 780095. This should ensure that we never
register a RenderFrameHost with the ProcessManager, which is not alive
and for which we may not subsequently receive a RenderFrameDeleted notification.

BUG=780095

Change-Id: I000080dd1a1288aa49246231a9466e2b50109e92
Reviewed-on: https://chromium-review.googlesource.com/835847
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Commit-Queue: Karan Bhatia <karandeepb@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#525455}(cherry picked from commit fd85aab1a25053e453e1ef74ecb1fc4ff0fa053e [formerly ad3b727a255305414d1cfeef0c8b2dfefa3220fd])
Reviewed-on: https://chromium-review.googlesource.com/879530
Reviewed-by: Karan Bhatia <karandeepb@chromium.org>
Cr-Commit-Position: refs/branch-heads/3282@{#572}
Cr-Branched-From: ab27e3b6feefb91d7474fe53634cf832758e747e [formerly 5fdc0fab22ce7efd32532ee989b223fa12f8171e]-refs/heads/master@{#520840}
Former-commit-id: 6caad66f6b6cc80dae9df49b63660cb5df7d2e8d