Reduce the amount of padding needed for SDFs.
SDF glyphs and paths were padded by 4 pixels, and had an inset of 2
pixels within that for rendering the shape. We should be able to
use a pad of 1, and the original bounds for the shape. This allows us
to increase the maximum size in the atlas.

Bug: skia:
Change-Id: I844d6afc4e0a52fc49fe1f6df94b635779290307
Reviewed-on: https://skia-review.googlesource.com/101440
Reviewed-by: Brian Salomon <bsalomon@google.com>
Commit-Queue: Jim Van Verth <jvanverth@google.com>

Former-commit-id: b42610aa9b0e431c5658f7f9ada8553e4f90869c