Reduce the amount of padding needed for SDFs, take 2
SDF glyphs and paths were padded by 4 pixels, and had an inset of 2
pixels within that for rendering the shape. We should be able to
use a pad of 2, and outset the original bounds by one texel. This allows us
to increase the maximum size in the atlas.

Change-Id: I06b229c7a3bc907d75d320461320d0b588f9a77c
Reviewed-on: https://skia-review.googlesource.com/101843
Reviewed-by: Brian Salomon <bsalomon@google.com>
Commit-Queue: Jim Van Verth <jvanverth@google.com>

Former-commit-id: 10105308dc86dccd3d4a0c8b9863d7c749440b21