Site isolation: report local size to content
If a CSS scale is applied to an OOPIF, the physical backing size was
incorrectly set. The entirety of the OOPIF is expected to be displayed even
if scaled. With this change, the scaling occurs entirely in the parent and the
child is oblivious to the scaling.

Note that RenderWidgetHostView::GetViewBounds is used in different ways:

1. To set the layout size of the child renderer.
2. To know position for popups.

For 1, it should be oblivious to transforms in the parent.
For 2, it needs to take into account transforms in the parent.

To fix this, we need to decouple position and size. This is a fairly major
refactor, so I'll leave it outside the scope of this hotfix for an immediate
site isolation problem.

TBR=fsamuel@chromium.org

(cherry picked from commit 1e31a010b9fb855de41b54a5923a290f6d2f959e [formerly 82fd52abc249b5535789a42d2ead47e45a44b580])

Bug: 806937
Change-Id: I8efeab97002983071aa02519dbfa8a82cf0a9ba3
Reviewed-on: https://chromium-review.googlesource.com/891663
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Antoine Labour <piman@chromium.org>
Commit-Queue: Fady Samuel <fsamuel@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#533403}
Reviewed-on: https://chromium-review.googlesource.com/907212
Reviewed-by: Fady Samuel <fsamuel@chromium.org>
Cr-Commit-Position: refs/branch-heads/3325@{#367}
Cr-Branched-From: 09f06fcf0fa239ed52c280c47686a4b3d4ed8f36 [formerly bc084a8b5afa3744a74927344e304c02ae54189f]-refs/heads/master@{#530369}
Former-commit-id: 67dfa6f581eacce483aa33c3325d6107cc3e5ef4