Remove event coalescing from RenderWidgetTargeter.
The code removed in this CL is believed to be the root cause of the
crash in the associated bug.

The mechanism is this: when events are coalesced, some event types still
expect all of the original events to be properly acked, and at present
RenderWidgetTargeter doesn't implement this. The crash, in particular,
is the result of touch events being acked in a way that the unique ids
don't line up with the queue in
TouchDispositionGestureFilter::OnTouchEventAck()
because some of the ids are missing after being coalesced.

At present we will only disable coalescing for touch and gesture
events,but that's reasonable since in most cases we only everhit-test
the initial TouchStart and everything else in the sequence latches to
its target.

Further, it's not obvious if there's a significant performance
benefit to the coalescing these events, so for now we remove it.
In any case, implementing proper acking of these events would result
in a patch too big to merge back to M65.

TBR=wjmaclean@chromium.org

(cherry picked from commit 6a67adaafb66fd50190b72e949f2e6858cfa3b97 [formerly 29de174eb52e75a7606f00b6fe3124a953b6d8c5])

Bug: 807507
Change-Id: I7bc45596d5a403118fd3905950fc777a8e4bfa1f
Reviewed-on: https://chromium-review.googlesource.com/913549
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: James MacLean <wjmaclean@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#536082}
Reviewed-on: https://chromium-review.googlesource.com/919964
Reviewed-by: James MacLean <wjmaclean@chromium.org>
Cr-Commit-Position: refs/branch-heads/3325@{#470}
Cr-Branched-From: 09f06fcf0fa239ed52c280c47686a4b3d4ed8f36 [formerly bc084a8b5afa3744a74927344e304c02ae54189f]-refs/heads/master@{#530369}
Former-commit-id: 1216cbf51b42c99033b67f10d72bcc72edb9c46b