Merge to M65: Fix critical accessibility bug causing page content to be inaccessible
In Blink accessibility, when an AXObject is "ignored", it's not
included in the accessibility tree. The ignored node's parent
includes the ignored node's children directly, leaving it out of the
tree.

As an example, consider this html:
<html><body><div><button/></div></body></html>

The current heuristics leave the body element out of the accessibility tree.
The accessibility tree looks like:

rootWebArea (the html node)
  genericContainer (the div)
    button

There are many ways for a node that's ignored to become unignored,
like adding an id or an ARIA attribute. Typically this fires the right
notifications and the node gets spliced into the accessibility tree.

The bug in this case is that adding a click listener to an ignored node
causes it to be not ignored anymore, but it does not correctly invalidate
its parent. In the example above, adding a click listener to the body and
then modifying the html element "breaks" the tree. The HTML node thinks
its child is the DIV, but the DIV thinks its parent is the BODY. This
inconsistency breaks the entire page's accessibility tree.

The fix is to catch changes to a node's ignored state and call ChildrenChanged.

Bug: 797695
Change-Id: Ib8b077e3ebb96ec1634a847db7477589cde74012
Reviewed-on: https://chromium-review.googlesource.com/910123
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#535734}(cherry picked from commit a01037419896a0797604ba432a9c3c242b235bef [formerly 7dc3a173d256cf3edbca00d8ca2680a4e4b32a97])
TBR: nektar@chromium.org
Reviewed-on: https://chromium-review.googlesource.com/922561
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/branch-heads/3325@{#476}
Cr-Branched-From: 09f06fcf0fa239ed52c280c47686a4b3d4ed8f36 [formerly bc084a8b5afa3744a74927344e304c02ae54189f]-refs/heads/master@{#530369}
Former-commit-id: 3390c604ea990b54d6000c16de59c85e7ea7e0ce