Allow screen capture over ARC for testimage builds
This is for automated testing. During testing, we need to automatically
click on approval in the permissions dialog. This allows Android to have
a mechanism by which it calls into Chrome and requests that it accepts
the permission. This is ONLY allowed when the build is of type
'testimage'. This then allows the automatic test to proceed. It also
disables showing the notification that screen capture is activate since
during Android testing that is also not expected to be part of the
screen.


Bug: b:38452042, b:73045468
Test: CTS tests pass now
Change-Id: I87d46b1d4859e3fe6b1b034e6a020ceafc9b7357
Reviewed-on: https://chromium-review.googlesource.com/909299
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Luis Hector Chavez <lhchavez@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#541686}(cherry picked from commit 2378ce82e01599707d2e58899afd4dbecfc0bd78 [formerly cc9e13435e525d014b20e70d3fa5b4e2b9d80356])
Reviewed-on: https://chromium-review.googlesource.com/963231
Reviewed-by: James Hawkins <jhawkins@chromium.org>
Cr-Commit-Position: refs/branch-heads/3359@{#247}
Cr-Branched-From: 0d720befa68dd312f58d12edcf8c90bd2cf17d54 [formerly 66afc5e5d10127546cc4b98b9117aff588b5e66b]-refs/heads/master@{#540276}
Former-commit-id: fe1d88f2575a9c6adf895c3fbe201bad3a91b96d