Make get_images_from_skps always return success
Previously, it would fail (and thus turn the bot red) if an image failed
to decode (e.g. with SkCodec::Result of kInvalidInput). This tends to
happen because the captured page had not fully downloaded yet. Stop
treating this an error. It may still be useful to test using SkCodec
just to be sure that we didn't crash, so keep running the test.

Change-Id: I1a0530fb82fa32d95c5f01d5131626745cf6c4f2
Reviewed-on: https://skia-review.googlesource.com/114582
Reviewed-by: Ravi Mistry <rmistry@google.com>
Reviewed-by: Mike Reed <reed@google.com>
Commit-Queue: Leon Scroggins <scroggo@google.com>

Former-commit-id: af18f22da6c53346a326dcb0c1b237e20ca44fe4