[merge to m65] gpu: Handle sync token waits with invalid order numbers ahead of time
Background: A sequence (also called stream) has multiple command
buffers, and runs tasks like Flush in the order in which they're
received. Therefore, sequences are effectively flush ordered internally,
but can run out of order with respect to other sequences.

SyncPointManager ensures that there are no sync token wait deadlocks by
assigning an order number to each IPC message. When a command buffer
waits on a sync token, we enqueue the wait, and the order number of the
waiting message (Flush) with the releasing command buffer's sequence
("release" sequence). If the releasing command buffer does not retire
the sync token, we will still retire it once we go past the waiting
order number in the release sequence, or if there are no more messages
in the release sequence. This is to handle malicious renderers sending
the browser sync tokens that aren't retired, and prevent the browser
from hanging forever. It should also prevent deadlocks incorrect client
logic.

With the scheduler, we prevent a sequence from running if its next
message has sync token waits that aren't resolved yet. We do this by
adding sync token dependency hints in the Flush message. This is to
avoid unnecessary context switching caused by executing a command buffer
just to encounter a wait immediately. However, these dependencies are
just hints, and we don't rely on them for correctness, i.e. it's
possible for the Flush message to contain no hints, but still have sync
token waits in the command buffer.

In this bug, we encounter a cyclic dependency graph between sequences,
where sequence A waits on sequence B with order number X which is less
than the order number of B's next task, and sequence B waits on sequence
A with an order number P which is greater than the order number of A's
next task (which is X). This is caused by a client error that we haven't
figured out yet, but we should still be resilient to it.

However, this does not work as expected because the scheduler
deschedules both sequences, and asks SyncPointManager to notify when the
waits are resolved. SyncPointManager cannot resolve cyclic dependencies
until one of the sequences runs. The logic to retire pending waits runs
either when you start or finish a message with order number X.
SyncPointManager doesn't know what the next order number in the task
queue is. This worked correctly when we didn't have sync token
dependency hints, and we would run one of the sequences in the cycle
first. In the above example, the wait from A to B is invalid because B
doesn't have a task with order number less than X. Only after starting
to run the next task can SyncPointManager know to retire A's pending
wait on B.

To fix this, SyncPointManager would need the queue of order numbers for
each sequence, and use that to invalidate any waits before the next
order number. In the above example, A's wait on B would be deemed
invalid because SyncPointManager can look ahead and see that the next
task order number on B is greater than X. Ironically, the scheduler has
this information in its tasks_ list, but doesn't use it because we
wanted sync point manager to be the source of truth to support both
scheduler and non-scheduler code paths over a year ago.

This has been a bug since M63. Maybe we didn't hit it because the client
started doing something incorrectly recently. One possibility is that
the client is sending sync tokens to another process, but without the
synchronous IPC to "verify" or confirm that an order number is assigned
before sending it to another process.

The short term fix is to add a queue of order numbers to
SyncPointOrderData which is SyncPointManager's state about a sequence.
Today, we have unprocessed_order_num_ which is the back of this queue,
processed_order_num_ which is the last order number popped off the
queue, but what we really want is the front of the queue (next task
order number) any time a wait is enqueued.

Long term, we may want to make the Scheduler the source of truth for
sync tokens.

Cq-Include-Trybots: luci.chromium.try:linux_optional_gpu_tests_rel;master.tryserver.chromium.android:android_optional_gpu_tests_rel;master.tryserver.chromium.mac:mac_optional_gpu_tests_rel;master.tryserver.chromium.win:win_optional_gpu_tests_rel
Change-Id: I87c2680b14ccb04b5ba58ab329db683c400f45b3
Bug: 797137
Reviewed-on: https://chromium-review.googlesource.com/969639
Commit-Queue: Sunny Sachanandani <sunnyps@chromium.org>
Reviewed-by: Antoine Labour <piman@chromium.org>
Reviewed-by: Victor Miura <vmiura@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#544267}(cherry picked from commit 75e71fd10d41060785b8e93ad0b7bac1031920a4 [formerly 3743d0f5f0614909a354a8df3af879f0eff21c42])
Reviewed-on: https://chromium-review.googlesource.com/970081
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Cr-Commit-Position: refs/branch-heads/3325@{#724}
Cr-Branched-From: 09f06fcf0fa239ed52c280c47686a4b3d4ed8f36 [formerly bc084a8b5afa3744a74927344e304c02ae54189f]-refs/heads/master@{#530369}
Former-commit-id: fd98c46928d599bcc351c3c8e772a202438a4fbf