Merge to M66: OOPIF: Transform point of resent gesture event instead of applying offset.
Note that this CL differs from the original as the representation of
coordinates in WebGestureEvent has changed between M66 and ToT.

OOPIF: Transform point of resent gesture event instead of applying offset.

When bubbling scroll events to ancestor views, we adjust the event's
position by applying the parent's offset. This is incorrect if an OOPIF
is subject to CSS scaling, since the point in the child is scaled. We
now use a transformation function which accounts for scaling.

We were also assuming that the event was being resent to the parent. This
is true for GestureScrollBegin events, but for the remaining events in the
sequence, we may bubble directly to an ancestor, so those events have the
coordinates for the wrong view. We now defer the transformation of the
coordinates to the target view's coordinate space to the
RenderWidgetHostInputEventRouter which knows which view is the target.

(cherry picked from commit a2dc7f4c481569da7fba28d1733bd936c48e9da7 [formerly 8961ab574285a3907e22c60eb78eda0d82904e50])

Bug: 626020, 817392
Change-Id: I86331f7bfc5fe7dd9e9032beb16ebb75b579f5b1
Reviewed-on: https://chromium-review.googlesource.com/978647
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Nasko Oskov <nasko@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#545890}
Reviewed-on: https://chromium-review.googlesource.com/983756
Cr-Commit-Position: refs/branch-heads/3359@{#496}
Cr-Branched-From: 0d720befa68dd312f58d12edcf8c90bd2cf17d54 [formerly 66afc5e5d10127546cc4b98b9117aff588b5e66b]-refs/heads/master@{#540276}
Former-commit-id: 5857d8c65e779b222549d5794f28941300952f5a