Fix renderer process termination when navigating to about:blank#foo. (M67 merge)
When a remote frame is navigated to about:blank#foo in --site-per-process mode
it is routed to the browser process through a RenderFrameProxyHost. It then
considers the navigation to be cross-site due to check for explicit
"about:blank" URL in DetermineSiteInstanceForURL. This causes a
cross-process navigation and Blink considers the about:blank#foo URL to be
same-document with the initial empty document. In turn, it sends a same
document commit to the browser process, which considers it a mismatch and
terminates the renderer.

This CL fixes DetermineSiteInstanceForURL to use GURL::IsAboutBlank, which
should cover various types of about:blank URLs. It also ensures that Blink
does not consider the first navigation in a provisional LocalFrame to be
eligible for same-document commit.

Bug: 833005
Change-Id: Ia7d849effbc52efb242b65ca822512bd44b016ff
Reviewed-on: https://chromium-review.googlesource.com/1031374
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Nasko Oskov <nasko@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#554436}(cherry picked from commit c5199977c984adb6376f7d7fee5f84499e85be65 [formerly cae49c8c95e1ca05abdb994a4dbb4d8400d437b6])
Reviewed-on: https://chromium-review.googlesource.com/1037647
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Cr-Commit-Position: refs/branch-heads/3396@{#420}
Cr-Branched-From: c765e9889451d875603d7b6fee6f8ccf953bf03f [formerly 9ef2aa869bc7bc0c089e255d698cca6e47d6b038]-refs/heads/master@{#550428}
Former-commit-id: d3319f11f005c53b447b562ebe3064582088e81b