Move close button to right side in Android horizontal tab switcher
Currently, in the new Android horizontal tab switcher, we show the close button
on the left side of the tab (in LTR mode) in both portrait and landscape
orientations. We've decided we want to put it on the right side to improve
readability of the favicons. This CL moves the close button to be on the right
in both portrait and landscape modes.

This CL also cleans up the logic for determining which side to put the button
on, since we currently repeat the same logic in multiple places, which is very
difficult to maintain. I'm adding a parameter on LayoutTab that Stack can keep
updated so TabListSceneLayer doesn't have to replicate the logic for the check.

Bug: 849397,831359
Change-Id: I01559d56d084ffe3c2a2a35468c8a62fa5a16a16
Reviewed-on: https://chromium-review.googlesource.com/1089572
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Ryan Landay <rlanday@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#565431}(cherry picked from commit 4f00de8b14af1a7389b821e1b97dc91a6b183e12 [formerly 01b7e11329b04b4abdd4181b6119fcd384c69d6b])
Reviewed-on: https://chromium-review.googlesource.com/1091878
Reviewed-by: Ryan Landay <rlanday@chromium.org>
Cr-Commit-Position: refs/branch-heads/3440@{#251}
Cr-Branched-From: dc83408e893a6ef8ceae89e5588f0a97ba95eaad [formerly 010ddcfda246975d194964ccf20038ebbdec6084]-refs/heads/master@{#561733}
Former-commit-id: 96907563eae68862d84d1ee549aedec6fb3194d7