Notify MediaStreamObserver about new stream when tracks start.
This CL makes sure the MediaStreamObserver object is updated
in the same task where tracks start. Failing to do this has caused
makes the observer and VideoCaptureImpl to have different views
of which devices are opened for a brief time during which it is
possible that an attempt to suspend a track that has not started
can be made, which results in a crash.

A test will be added in a follow-up CL since it requires changes that
are much bigger than the fix, which would make merging more difficult.

Bug: 847367
Change-Id: Id88372d1ab5d03d6cd55f902ac8901ff6bd7bdea
Reviewed-on: https://chromium-review.googlesource.com/1087274
Reviewed-by: Christian Fremerey <chfremer@chromium.org>
Commit-Queue: Guido Urdaneta <guidou@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#564592}(cherry picked from commit 9b5735951c9b2dde59c80233f79322eb9fe3c388 [formerly 09b695b523127b203ccb6c6fc053da1153c017dc])
Reviewed-on: https://chromium-review.googlesource.com/1093394
Cr-Commit-Position: refs/branch-heads/3440@{#261}
Cr-Branched-From: dc83408e893a6ef8ceae89e5588f0a97ba95eaad [formerly 010ddcfda246975d194964ccf20038ebbdec6084]-refs/heads/master@{#561733}
Former-commit-id: f8e2df3fec3c8c8093e7fd5ace2f8178ce6453c5