Execute plugin notifications synchronously (#32)
If the 'bbOnUpdateGeometry' and 'bbOnUpdateVisiblity' properties of the plugin
have JavaScript functions assigned to them, then, the plugin will execute those
functions immediately (in addition to continuing to dispatch DOM events to
perform the notification).

Since JavaScript execution is normally forbidden during layout, the callbacks
are invoked by passing them to a new function 'blink::WebScriptBindings',
which uses blink's 'ScriptForbiddenScope::AllowUserAgentScript' class to
establish a scope that allows such scripts to run.

The purpose of these changes is to allow plugin resize handling to take place
synchronously with respect to the Windows message that caused the DOM layout
in the first place. A very common use case for this plugin object is to
change the geometry of a child HWND to be that of a DOM element, and executing
that operation synchronously appears to improve performance because it allows
the operating system to coalesce rapid move and size events.